Separated Command to and interface and utils class, also modifying it's usage a bit. Added final to parse method.
